fix(ci): resolve all CI failures
- contracts: remove redundant u32 casts caught by clippy (storage.rs in
  credit-registry, retirement-tracker, marketplace)
- api: fix postgres health-check --health-cmd quoting in CI workflow
- api/indexer: add eslint + @typescript-eslint deps and .eslintrc.js configs
  (no-unused-vars/no-var-requires as warn; stubs are intentional)
- deps: add eslint@8, @typescript-eslint/parser@7, @typescript-eslint/eslint-plugin@7

+++ contracts/credit-registry/src/storage.rs
@@ -136,8 +136,8 @@ pub fn read_credits_by_issuer(env: &Env, issuer: &Address, offset: u32, limit: u
         .instance()
         .get(&(ISSUER_LIST_PREFIX, issuer.clone()))
         .unwrap_or_else(|| vec![env]);
-    let end = core::cmp::min(offset.saturating_add(limit), list.len() as u32);
-    if offset >= list.len() as u32 {
+    let end = core::cmp::min(offset.saturating_add(limit), list.len());
+    if offset >= list.len() {
         return vec![env];
     }
     let mut result = vec![env];
